introduceRemoveDialogUncontrolled

+++ faktorio-fe/src/components/RemoveDialog.tsx
@@ -10,9 +10,9 @@ import {
 } from './ui/dialog'
 import { Button } from './ui/button'
 import { Spinner } from './ui/spinner'
-import ReactDOM from 'react-dom'
+import ReactDOM, { Root } from 'react-dom/client'
 
-export function RemoveDialog({
+export function RemoveDialogInner({
   open,
   onCancel,
   onRemove,
@@ -21,15 +21,15 @@ export function RemoveDialog({
   loading
 }: {
   open: boolean
-  onCancel: (open: boolean) => void
+  onCancel: () => void
   onRemove: () => void
-  title: string
-  description: React.ReactNode
+  title: React.ReactNode
+  description?: React.ReactNode
   loading?: boolean
 }) {
   return (
     <Dialog open={open} onOpenChange={!loading ? onCancel : undefined}>
-      <DialogContent className="sm:max-w-[425px]" close={!loading}>
+      <DialogContent className="sm:max-w-[425px]">
         <DialogHeader>
           <DialogTitle>{title}</DialogTitle>
           <DialogDescription className="pt-2">{description}</DialogDescription>
@@ -38,7 +38,7 @@ export function RemoveDialog({
         <DialogFooter className="gap-2">
           <Button
             variant="outline"
-            onClick={() => onCancel(false)}
+            onClick={() => onCancel()}
             disabled={loading}
           >
             Zrušit
@@ -65,46 +65,60 @@ export function RemoveDialogUncontrolled({
   description,
   children,
   loading
+}: {
+  onRemove: () => void
+  title: React.ReactNode
+  description?: React.ReactNode
+  children: React.ReactNode
+  loading?: boolean
 }) {
   const [open, setOpen] = useState(false)
 
+  const root = React.useRef<Root | null>(null)
   // Create a state for the portal div element
-  const [portalDiv, setPortalDiv] = useState(null)
+  // const [portalDiv, setPortalDiv] = useState(null)
 
   // Initialize the portal div on mount and clean up on unmount
   useEffect(() => {
     // Create a new div that will be the portal for the RemoveDialog
-    const div = document.createElement('div')
-    const portalRoot = document.getElementById('portal-root')
-    portalRoot.appendChild(div)
-    setPortalDiv(div)
+
+    if (!root.current) {
+      root.current = ReactDOM.createRoot(
+        document.getElementById('portal-root') // This is the target for the portal
+      )
+    }
+    if (!root.current) {
+      throw new Error('root.current is null')
+    }
+    root.current.render(
+      <RemoveDialogInner
+        open={open}
+        onCancel={() => {
+          root.current?.unmount()
+        }}
+        onRemove={() => {
+          onRemove()
+          root.current?.unmount()
+        }}
+        title={title}
+        description={description}
+        loading={loading}
+      />
+    )
+
+    // setPortalDiv(div)
 
     // Cleanup function to remove the div from the body when the component unmounts
     return () => {
       // portalRoot.removeChild(div)
     }
-  }, [])
+  }, [open])
 
   return (
     <>
-      <div className="flex" onClick={() => setOpen(true)}>
+      <div className="w-full" onClick={() => setOpen(true)}>
         {children}
       </div>
-      {portalDiv &&
-        ReactDOM.createPortal(
-          <RemoveDialog
-            open={open}
-            onCancel={() => setOpen(false)}
-            onRemove={() => {
-              onRemove()
-              setOpen(false) // Optionally close the dialog after removal
-            }}
-            title={title}
-            description={description}
-            loading={loading}
-          />,
-          portalDiv // This is the target for the portal
-        )}
     </>
   )
 }
